* trunk: (590 commits)
  [FIX] renamed method II
  [FIX] renamed method
  [FIX] declaration of PlusJob
  Workflow Engine: Checked directory existence
  Mail: Respected other image mime types
  fixed bug #21673: Failed test: Custom metadata does not reliably survive Export
  partly fixed 21971: Failed test: Bestehendes Wiki mit benutzerdefinierten Metadaten importieren
  fixed exception if importer not found
  fixed bug #19200: Failed test: Ansicht Kompetenz-Ergebnisse
  [Fix 21636] Day/Week view recurrences
  fixed bug #21631: Failed test: Fragenübersicht einfügen- Preview of SCO doesn't work
  Calendar modal ps reordering + fixes.
  fixed sco preview
  [OrgUnits] employee and superior positions will be filled with the employee and superior roles of previous ilias version.
  User Gallery: Fixed #22219
  User Gallery: Fixed #22219
  fixed bug #21822: Failed test: Mehrsprachiges Lernmodul exportieren
  Modified test template for phpunit test
  remove public api of OrgUnits for adding staff to local roles.
  [Fix 21524] cal. list view: strip tags for description and property values
  ...

# Conflicts:
#	src/UI/Implementation/Component/ComponentHelper.php
#	templates/default/less/Services/Form/delos.less
#	templates/default/less/variables.less

+++ Modules/Survey/classes/class.ilObjSurvey.php
@@ -595,30 +595,68 @@ class ilObjSurvey extends ilObject
 	function insertQuestion($question_id) 
 	{
 		$ilDB = $this->db;
-		
+
+		$this->log->debug("insert question, id:".$question_id);
+
 		include_once "./Modules/SurveyQuestionPool/classes/class.SurveyQuestion.php";
 		if (!SurveyQuestion::_isComplete($question_id))
 		{
+			$this->log->debug("question is not complete");
 			return FALSE;
 		}
 		else
 		{
 			// get maximum sequence index in test
+			// @todo: refactor this
 			$result = $ilDB->queryF("SELECT survey_question_id FROM svy_svy_qst WHERE survey_fi = %s",
 				array('integer'),
 				array($this->getSurveyId())
 			);
 			$sequence = $result->numRows();
 			$duplicate_id = $this->duplicateQuestionForSurvey($question_id);
+			$this->log->debug("duplicate, id: ".$question_id.", duplicate id: ".$duplicate_id);
+
+			// check if question is not already in the survey, see #22018
+			if ($this->isQuestionInSurvey($duplicate_id))
+			{
+				return false;
+			}
+
 			$next_id = $ilDB->nextId('svy_svy_qst');
 			$affectedRows = $ilDB->manipulateF("INSERT INTO svy_svy_qst (survey_question_id, survey_fi, question_fi, sequence, tstamp) VALUES (%s, %s, %s, %s, %s)",
 				array('integer', 'integer', 'integer', 'integer', 'integer'),
 				array($next_id, $this->getSurveyId(), $duplicate_id, $sequence, time())
 			);
+
+			$this->log->debug("added entry to svy_svy_qst, id: ".$next_id.", question id: ".$duplicate_id.", sequence: ".$sequence);
+
 			$this->loadQuestionsFromDb();
 			return TRUE;
 		}
 	}
+
+	/**
+	 * Check if a question is already in the survey
+	 *
+	 * @param question id (as primary key from svy_question table)
+	 * @return bool
+	 */
+	function isQuestionInSurvey($a_question_fi)
+	{
+		global $DIC;
+//return false;
+		$ilDB = $DIC->database();
+
+		$set = $ilDB->query("SELECT * FROM svy_svy_qst ".
+			" WHERE survey_fi = ".$ilDB->quote($this->getSurveyId(), "integer").
+			" AND question_fi = ".$ilDB->quote($a_question_fi, "integer"));
+		if ($rec = $ilDB->fetchAssoc($set))
+		{
+			return true;
+		}
+		return false;
+	}
+
 
 
 /**
@@ -853,7 +891,9 @@ class ilObjSurvey extends ilObject
 	function saveQuestionsToDb() 
 	{
 		$ilDB = $this->db;
-		
+
+		$this->log->debug("save questions");
+
 		// gather old questions state
 		$old_questions = array();
 		$result = $ilDB->queryF("SELECT survey_question_id,question_fi,sequence".
@@ -863,35 +903,36 @@ class ilObjSurvey extends ilObject
 		);		
 		while($row = $ilDB->fetchAssoc($result))
 		{
-			$old_questions[$row["question_fi"]] = $row;
+			$old_questions[$row["question_fi"]] = $row;		// problem, as soon as duplicates exist, they will be hidden here
 		}		
 		
 		// #15231 - diff with current questions state
 		$insert = $update = $delete = array();		
 		foreach($this->questions as $seq => $fi)
 		{
-			if(!array_key_exists($fi, $old_questions))
+			if(!array_key_exists($fi, $old_questions))		// really new fi IDs
 			{
-				$insert[] = $fi;
+				$insert[] = $fi;							// this should be ok, should not create duplicates here
  			}
-			else if($old_questions[$fi]["sequence"] != $seq)
+			else if($old_questions[$fi]["sequence"] != $seq)				// we are updating one of the duplicates (if any)
 			{
 				$update[$fi] = $old_questions[$fi]["survey_question_id"];
 			}
 			// keep track of still relevant questions
-			unset($old_questions[$fi]);
+			unset($old_questions[$fi]);						// deleting old question, if they are not in current array
 		}		
 		
 		// delete obsolete question relations
 		if(sizeof($old_questions))
 		{
 			$del_ids = array();
-			foreach($old_questions as $old)
+			foreach($old_questions as $fi => $old)
 			{
 				$del_ids[] = $old["survey_question_id"];
-			}			
-			$ilDB->manipulate("DELETE FROM svy_svy_qst".
-				" WHERE ".$ilDB->in("survey_question_id", $del_ids, "", "integer"));			
+			}
+			$ilDB->manipulate($q = "DELETE FROM svy_svy_qst".
+				" WHERE ".$ilDB->in("survey_question_id", $del_ids, "", "integer"));
+			$this->log->debug("delete: ".$q);
 		}
 		unset($old_questions);
 		
@@ -900,20 +941,26 @@ class ilObjSurvey extends ilObject
 		{
 			if(in_array($fi, $insert))
 			{
-				$next_id = $ilDB->nextId('svy_svy_qst');
-				$ilDB->manipulateF("INSERT INTO svy_svy_qst".
-					" (survey_question_id, survey_fi, question_fi, heading, sequence, tstamp)".
-					" VALUES (%s, %s, %s, %s, %s, %s)",
-					array('integer','integer','integer','text','integer','integer'),
-					array($next_id, $this->getSurveyId(), $fi, NULL, $seq, time())
-				);
+				// check if question is not already in the survey, see #22018
+				if (!$this->isQuestionInSurvey($fi))
+				{
+					$next_id = $ilDB->nextId('svy_svy_qst');
+					$ilDB->manipulateF("INSERT INTO svy_svy_qst" .
+						" (survey_question_id, survey_fi, question_fi, heading, sequence, tstamp)" .
+						" VALUES (%s, %s, %s, %s, %s, %s)",
+						array('integer', 'integer', 'integer', 'text', 'integer', 'integer'),
+						array($next_id, $this->getSurveyId(), $fi, null, $seq, time())
+					);
+					$this->log->debug("insert svy_svy_qst, id:" . $next_id . ", fi: " . $fi . ", seq:" . $seq);
+				}
 			}
 			else if(array_key_exists($fi, $update))
 			{
 				$ilDB->manipulate("UPDATE svy_svy_qst".
 					" SET sequence = ".$ilDB->quote($seq, "integer").
 					", tstamp = ".$ilDB->quote(time(), "integer").
-					" WHERE survey_question_id = ".$ilDB->quote($update[$fi], "integer"));				
+					" WHERE survey_question_id = ".$ilDB->quote($update[$fi], "integer"));
+				$this->log->debug("update svy_svy_qst, id:".$update[$fi].", fi: ".$fi.", seq:".$seq);
 			}
 		}
 	}
@@ -1151,6 +1198,65 @@ class ilObjSurvey extends ilObject
 			$this->questions[$data["sequence"]] = $data["question_fi"];
 		}
 	}
+
+	/**
+	 * Remove duplicate sequence entries, see #22018
+	 */
+	function fixSequenceStructure()
+	{
+		global $DIC;
+
+		$ilDB = $DIC->database();
+//return;
+		// we keep all survey question ids with their lowest sequence
+		$result = $ilDB->queryF("SELECT * FROM svy_svy_qst WHERE survey_fi = %s ORDER BY sequence",
+			array('integer'),
+			array($this->getSurveyId())
+		);
+
+		// step 1: find duplicates -> $to_delete_ids
+		$fis = array();
+		$to_delete_ids = array();
+		while ($data = $ilDB->fetchAssoc($result))
+		{
+			if (in_array($data["question_fi"], $fis))		// found a duplicate
+			{
+				$to_delete_ids[] = $data["survey_question_id"];
+			}
+			else
+			{
+				$fis[] = $data["question_fi"];
+			}
+		}
+
+		// step 2: we delete the duplicates
+		if (count($to_delete_ids) > 0)
+		{
+			$ilDB->manipulate($q = "DELETE FROM svy_svy_qst" .
+				" WHERE " . $ilDB->in("survey_question_id", $to_delete_ids, false, "integer") .
+				" AND survey_fi = " . $ilDB->quote($this->getSurveyId(), "integer"));
+			$this->log->debug("delete: " . $q);
+
+			$ilDB->manipulate($q = "DELETE FROM svy_qblk_qst " .
+				" WHERE " . $ilDB->in("question_fi", $fis, true, "integer") .
+				" AND survey_fi = " . $ilDB->quote($this->getSurveyId(), "integer"));
+			$this->log->debug("delete: " . $q);
+		}
+
+		// step 3: we fix the sequence
+		$set = $ilDB->query("SELECT * FROM svy_svy_qst ".
+			" WHERE survey_fi = " . $ilDB->quote($this->getSurveyId(), "integer")." ORDER BY sequence");
+		$seq = 0;
+		while ($rec = $ilDB->fetchAssoc($set))
+		{
+			$ilDB->manipulate($q = "UPDATE svy_svy_qst SET ".
+				" sequence = ".$ilDB->quote($seq++ , "integer").
+				" WHERE survey_question_id = ".$ilDB->quote($rec["survey_question_id"], "integer")
+			);
+			$this->log->debug("update: " . $q);
+		}
+	}
+
 
 /**
 * Sets the authors name of the ilObjSurvey object
@@ -1918,16 +2024,42 @@ class ilObjSurvey extends ilObject
 	{
 		$ilDB = $this->db;
 
+		// see #22018
+		if (!$this->isQuestionInAnyBlock($question_id))
+		{
+			$next_id = $ilDB->nextId('svy_qblk_qst');
+			$affectedRows = $ilDB->manipulateF("INSERT INTO svy_qblk_qst (qblk_qst_id, survey_fi, questionblock_fi, " .
+				"question_fi) VALUES (%s, %s, %s, %s)",
+				array('integer', 'integer', 'integer', 'integer'),
+				array($next_id, $this->getSurveyId(), $questionblock_id, $question_id)
+			);
 
-		$next_id = $ilDB->nextId('svy_qblk_qst');
-		$affectedRows = $ilDB->manipulateF("INSERT INTO svy_qblk_qst (qblk_qst_id, survey_fi, questionblock_fi, " .
-			"question_fi) VALUES (%s, %s, %s, %s)",
-			array('integer','integer','integer','integer'),
-			array($next_id, $this->getSurveyId(), $questionblock_id, $question_id)
-		);
-		
-		$this->deleteConstraints($question_id); // #13713
+			$this->deleteConstraints($question_id); // #13713
+		}
 	}
+
+	/**
+	 * Is question already in a block?
+	 *
+	 * @param int $a_question_fi question id as in svy_question
+	 * @return bool
+	 */
+	function isQuestionInAnyBlock($a_question_fi)
+	{
+		global $DIC;
+
+		$ilDB = $DIC->database();
+
+		$set = $ilDB->query("SELECT * FROM svy_qblk_qst ".
+			" WHERE survey_fi = ".$ilDB->quote($this->getSurveyId(), "integer").
+			" AND question_fi = ".$ilDB->quote($a_question_fi, "integer"));
+		if ($rec = $ilDB->fetchAssoc($set))
+		{
+			return true;
+		}
+		return false;
+	}
+
 
 /**
 * Returns the question titles of all questions of a question block
@@ -1976,7 +2108,10 @@ class ilObjSurvey extends ilObject
 	function &getQuestionblockQuestionIds($questionblock_id)
 	{
 		$ilDB = $this->db;
-		$result = $ilDB->queryF("SELECT question_fi FROM svy_qblk_qst WHERE questionblock_fi = %s",
+
+		// we need a correct order here, see #22011
+		$result = $ilDB->queryF("SELECT a.question_fi FROM svy_qblk_qst a JOIN svy_svy_qst b ON (a.question_fi = b.question_fi) ".
+			" WHERE a.questionblock_fi = %s ORDER BY b.sequence",
 			array("integer"),
 			array($questionblock_id)
 		);
@@ -1985,9 +2120,13 @@ class ilObjSurvey extends ilObject
 		{
 			while ($data = $ilDB->fetchAssoc($result))
 			{
-				array_push($ids, $data['question_fi']);
+				if (!in_array($data['question_fi'], $ids))		// no duplicates, see #22018
+				{
+					array_push($ids, $data['question_fi']);
+				}
 			}
 		}
+
 		return $ids;
 	}
 	
@@ -2062,13 +2201,16 @@ class ilObjSurvey extends ilObject
 			$questionblock_id = $next_id;
 			foreach ($questions as $index)
 			{
-				$next_id = $ilDB->nextId('svy_qblk_qst');
-				$affectedRows = $ilDB->manipulateF("INSERT INTO svy_qblk_qst (qblk_qst_id, survey_fi, questionblock_fi, " .
-					"question_fi) VALUES (%s, %s, %s, %s)",
-					array('integer','integer','integer','integer'),
-					array($next_id, $this->getSurveyId(), $questionblock_id, $index)
-				);
-				$this->deleteConstraints($index);
+				if (!$this->isQuestionInAnyBlock($index))
+				{
+					$next_id = $ilDB->nextId('svy_qblk_qst');	// #22018
+					$affectedRows = $ilDB->manipulateF("INSERT INTO svy_qblk_qst (qblk_qst_id, survey_fi, questionblock_fi, " .
+						"question_fi) VALUES (%s, %s, %s, %s)",
+						array('integer', 'integer', 'integer', 'integer'),
+						array($next_id, $this->getSurveyId(), $questionblock_id, $index)
+					);
+					$this->deleteConstraints($index);
+				}
 			}
 		}
 	}
@@ -2349,6 +2491,7 @@ class ilObjSurvey extends ilObject
 			}
 			$counter++;
 		}
+
 		return $all_pages;
 	}
 	
@@ -4537,8 +4680,8 @@ class ilObjSurvey extends ilObject
 						array(
 							"accesscode" => $data["code"],
 							"lang" => $lang
-						));				
-					$messagetext = str_replace('[url]', "<" . $url . ">", $messagetext);
+						));
+					$messagetext = str_replace('[url]', $url, $messagetext);
 					foreach ($data as $key => $value)
 					{
 						$messagetext = str_replace('[' . $key . ']', $value, $messagetext);
@@ -5540,6 +5683,7 @@ class ilObjSurvey extends ilObject
 			else
 			{			
 				$name = ilObjUser::_lookupName($row["user_id"]);
+				$name["name"] = $name["lastname"].", ".$name["firstname"];
 				$name["user_id"] = "u".$name["user_id"];
 				$name["email"] = ilObjUser::_lookupEmail($row["user_id"]);
 				$name["sent"] = $row["mail_sent"];
@@ -5558,7 +5702,8 @@ class ilObjSurvey extends ilObject
 					$res["a".$item["id"]] = array(
 						"user_id" => "a".$item["id"],
 						"lastname" => $item["last_name"],
-						"firstname" => $item["first_name"],						
+						"firstname" => $item["first_name"],
+						"name" => $item["last_name"].", ".$item["first_name"],
 						"login" => "",
 						"email" => $item["email"],
 						"code" => $item["code"],
